requirements.txt
discord.py==2.4.0
python-dotenv==1.0.0
openai==1.61.1
aiohttp==3.11.12
PyNaCl==1.5.0
SpeechRecognition==3.10.1
gTTS==2.5.1
sounddevice==0.4.6
vosk==0.3.44
websockets==12.0
numpy==1.26.4 